// apps/ships/src/app/api/entities/Starship.ts
export interface Starship {
    id: string;
    name: string;
    passengers: number;
    crew: string;
    details: string;
    MGLT: number;
    cargo_capacity: number;
    consumables: string;
    hyperdrive_rating: string;
    max_atmosphering_speed: number;
    model: string;
    starship_class: string;
}